project structure and refactored ride/get-by-id functions
