+++ Backend/BOSS/api/functions/user_functions.py
@@ -1,10 +1,6 @@
-from sched import scheduler
-
-from django.core.exceptions import ObjectDoesNotExist
-from django.utils import timezone
 from django.contrib.auth.hashers import make_password
-from .models import User, Ride, Notification, Report, Van
-import time
+from django.core.exceptions import ObjectDoesNotExist
+from api.models import User
 
 
 class UserFunctions:
@@ -169,144 +165,3 @@ class UserFunctions:
             return_list.append(temp_dic)
         """Return the list of users"""
         return return_list
-
-
-class RideManagement:
-    def create(self, rider: User, ride_info: dict) -> bool:
-        """Create a new ride."""
-        if rider.user_type != 'R':
-            return False
-
-        ride = Ride(
-            rider=rider,
-            pickup_location=ride_info.get('pickup_location'),
-            dropoff_location=ride_info.get('dropoff_location'),
-            num_passengers=ride_info.get('num_passengers', 1),
-            ADA_required=ride_info.get('ADA_required'),
-            van=ride_info.get('van'),
-        )
-        status = ride_info.get('status')
-        if status is not None:
-            ride.status = status
-        ride.save()
-        return True
-
-    def edit(self, ride_id: int, ride_info: dict) -> bool:
-        """Edit an existing ride."""
-        try:
-            ride = Ride.objects.get(id=ride_id)
-            for attr, value in ride_info.items():
-                setattr(ride, attr, value)
-            ride.save()
-            return True
-        except Ride.DoesNotExist:
-            return False
-
-    def delete(self, ride_id: int) -> bool:
-        """Delete a ride."""
-        try:
-            ride = Ride.objects.get(id=ride_id)
-            ride.delete()
-            return True
-        except Ride.DoesNotExist:
-            return False
-
-    def assign_driver(self, ride_id) -> bool:
-        """Assign a driver to an existing ride."""
-        try:
-            ride = self.get_by_id(ride_id)
-            if ride.driver is not None:
-                return False
-        except Ride.DoesNotExist:
-            return False
-
-        available_drivers = User.objects.filter(user_type='D', status__iexact='Available')
-        if not available_drivers.exists():
-            return False
-
-        driver = available_drivers.first()  # Choose the first available driver
-        ride.driver = driver
-        ride.van = VanManagement().get_by_driver_id(driver.id)
-        ride.pickup_time = timezone.now()
-        ride.status = 'Assigned'
-        ride.save()
-        driver.status = 'Assigned'
-        driver.save()
-
-        return True
-
-    def assign_driver_task(self, ride_id) -> bool:
-        result = self.assign_driver(ride_id)
-        retries = 0
-
-        while not result:
-            if retries > 10:
-                return False
-            time.sleep(60)
-            result = self.assign_driver(ride_id)
-        return result
-
-    def get_by_id(self, ride_id: int) -> Ride:
-        try:
-            ride = Ride.objects.get(id=ride_id)
-            return ride
-        except Ride.DoesNotExist:
-            return Ride()
-
-    def get_by_rider_id(self, _id: int, status: str = None) -> list:
-        try:
-            if status is not None:
-                return Ride.objects.filter(rider__id=_id, status=status)
-            return Ride.objects.filter(rider__id=_id)
-        except Ride.DoesNotExist:
-            return []
-
-    def get_by_driver_id(self, _id: int, status: str = None) -> list:
-        try:
-            if status is not None:
-                return Ride.objects.filter(driver__id=_id, status=status)
-            return Ride.objects.filter(driver__id=_id)
-        except Ride.DoesNotExist:
-            return []
-
-    def get_by_van_id(self, _id: int, status: str = None) -> list:
-        try:
-            if status is not None:
-                return Ride.objects.filter(van__id=_id, status=status)
-            return Ride.objects.filter(van__id=_id)
-        except Ride.DoesNotExist:
-            return []
-
-    def get_all(self) -> list:
-        """Get all rides."""
-        return list(Ride.objects.all())
-
-
-class NotificationManager:
-    def get(self, query: str, identity: str) -> Notification:
-        try:
-            return Notification.objects.get(id=query, rider=identity)
-        except Notification.DoesNotExist:
-            return None
-
-    def get_all(self) -> list:
-        return list(Notification.objects.all())
-
-
-class ReportManager:
-    def get(self, query: str, identity: str) -> Report:
-        try:
-            return Report.objects.get(id=query, reporter=identity)
-        except Report.DoesNotExist:
-            return []
-
-    def get_all(self) -> list:
-        return list(Report.objects.all())
-
-
-class VanManagement:
-    def get_by_id(self, _id: int) -> Van:
-        return Van.objects.get(id=_id)
-
-    def get_by_driver_id(self, _id: int) -> Van:
-        return Van.objects.get(driver__id=_id)
