+++ Backend/BOSS/api/functions/van_functions.py
@@ -1,0 +1,9 @@
+from api.models import Van
+
+
+class VanManagement:
+    def get_by_id(self, _id: int) -> Van:
+        return Van.objects.get(id=_id)
+
+    def get_by_driver_id(self, _id: int) -> Van:
+        return Van.objects.get(driver__id=_id)
