+++ Backend/BOSS/api/tests/__init__.py
@@ -1,0 +1,2 @@
+from .user_tests import UserFunctionsTests
+from .ride_tests import RideManagementTests
