+++ Backend/BOSS/api/views/__init__.py
@@ -1,0 +1,22 @@
+from .ride_views import (
+    get_all_rides,
+    get_ride_by_id,
+    get_ride_by_driver,
+    get_ride_by_rider,
+    get_ride_by_van,
+    assign_driver,
+    delete_ride,
+)
+
+from .users_view import (
+    ManageUsersView,
+)
+
+from .login_views import (
+    LoginView,
+    LogoutView,
+)
+
+from .register_views import (
+    RegisterView,
+)
