+++ Backend/BOSS/api/views/register_views.py
@@ -1,0 +1,13 @@
+from rest_framework import status
+from rest_framework.response import Response
+from rest_framework.views import APIView
+from api.serializer import UserSerializer
+
+
+class RegisterView(APIView):
+    def post(self, request):
+        serializer = UserSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response({"message": "Account created successfully"}, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
